Borg Jetpack Overhaul (#341)
Somewhat rewrote the cyborg jetpack to be an internal component instead of a software upgrade. this allows it to be removed or destroyed, and to work without putting something in the borg's inventory. Also fixed it, it was broken. Cyborg jetpack can now be installed on a wider array of borgs that have reason to go EVA: Crisis, security, engineering, construction (also the unused combat borg, incase that gets re-enabled)