Welding Tool tweaks (#517)
Welding tools can now be used to burn paper
The upgraded and experimental welding tools will now fit in a toolbelt. Upgraded renamed to advanced.
Fixed and implemented the Experimental Welding Tool, which has a regenerating fuel supply. Can be produced in R&D, requires 4 research in engineering and materials.