Bot Access (#630)
Added Master access to bot control panels for roboticists and research directors

Also added a return 1 to all bot construction steps, to prevent calling afterattack for the items used on the bot. This fixes a couple runtime errors and prevents more in future. Not included in changelog as this is irrelevant to end-users since they cant see runtime errors