Buckle Tweak
changes:
  - bugfix: "You can no longer buckle people into a chair/bed/etc which is already occupied."
  - tweak: "Buckling yourself into something is no longer a visible action, only you will see the message."

The former change, while working as intended, is causing undesireable behaviour, and i think this way makes better sense.

The last one is a brief idea i had to reduce noise in the chatlog. If anyone wants to see if you're buckled in, they can examine you. A message in the chatlog is needless noise and not relevant information to most people. It's most noticeable on the transfer shuttle where the large number of people buckling can impede communication