Table and Crate interactions (#1353)
Inspired partially by complaints about my changes to crates that made them hard to shoot open. This PR adds two main features which interact:

Crates can now be slid under tables, making them easier to stow away. A crate only fits under a table while closed, and can't be opened while under there.

Crates can now be hoisted up ontop of a table, this takes time depending on the mass of the crate. A crate ontop of a table works as normal, and it will also always block shots, making a table the ideal platform for firing an emitter at one.

A table can have a crate both under and ontop of it, allowing two crates to be stored on one tile, at some cost of convenience and effort. The crates have their pixel offsets modified so that both of them can be seen at once.