Tip of the Round (#1625)
Pretty simple. 10 seconds before the round begins, the game will choose a tip and display it.
Tips are handled in the config.
Admins also have a command to display their own customized tip, and also a command to display a pregenerated tip.
Both commands can be ran at any point in the run to good effect. The customized tip will replace the default, Tip of the round that procs 10 seconds before round-start tip, if it is run during the lobby.