Lawgiver and random fix (#1658)
Fixes an issue that popped up with obj/randoms, they werent getting sucked into lockers

Also fixes that lawgiver runtime.
The issue that previous fixes miss is that DM casts implicitly and without safety checks. Casting an animal to a human does not result in null. an istype check is needed