SolRep Uniforms (#1856)
After discussing it with Jackboot, I have decided to give the SolRep their own uniform. In an attempt to differentiate them from the captain, I created a new uniform sprite based off of the US Navy Dress Blue uniform.

Since SolRep's and their alternate title of IAA are very different, I decided it would be best to spawn all of the SA gear on the SolRep specifically, so that the IAA doesn't get greedy with a locker.

The SolRep now spawns with:
A regular pair of sunglasses.
A Lieutenant Sol Navy dress uniform
A Navy cap
Laced dress shoes
A ceremonial sword

They no longer spawn with a briefcase.

I also switched the sprite of the Marine dress cap and the Navy dress cap, because I believe that they better match their source this way.

I also renamed the dress whites uniform to be Sol Navy Commander Dress Uniform, and the Dress blues are Sol Navy Lieutenant Dress Uniform. This is to me the best way to convey rank without creating a whole new insignia accessory.