Nooses (#1805)
-ported nooses from hippie

They can be build with cable, if you stand on any table or bed. You can tie people, but it takes around 20 secs, and people that are not cuffed or disabled somehow can untie themselves. Also, it will produce some flavor texts while they are being hanged or a body is hanging from one.