Emergency Cooking Fixes (#1944)
This PR fixes a number of isssues with cooking that were overlooked during testing.

The cause of the infinite loop:
I neglected to add a buffer to hold cooked items to the microwave when i did likewise for other appliances, this is necessary for bulk cooking as it ensures that things which were just cooked, aren't considered as potential ingredients.
The lack of this buffer caused donk pockets to iterate forever, because the result (a cooked donk pocket) is a valid ingredient to create a cooked donk pocket. so it made the same recipe millions of times over. This is now fixed.

Runtime errors.
A few items would throw runtime errors because they override the make_food function which i significantly altered, and return data in the wrong format. This is now fixed. Affected items were: Donk pockets, valid salad, fortune cookies, amanita jelly.

Uncookable items.
A few recipes were impossible to create in appliances due to my neglecting support for adding non-snack items to cooking containers. This mostly affects diona roast and fortune cookies.. For now i've gone with a simple whitelist approach, allowing paper and holders to be added.