Cooking and Reagent Fixes(#1953)
Went searching for the cause of the reagent runtime issues. I don't think i found it yet, but i found and fixed a couple things that could be causing it.

Most notable is an egregious bug that caused animals' reagent holder to be nulled after they're dropped from a holder.
Also adds a little safety check to recipes which may fix it.

Fixes two cooking bugs as well. One that was allowing extra containers to be inserted into an already-full oven (they nested inside other containers, it was messy) and a tiny issue with batter overlays being incorrect on cooked meats.