[Newmap] Several modifications. (#2314)
Places the trash compactor in a pit, solving the trash distribution
issues and the maintenance access to the base of the pistons.

Moves the tables of the Security checkpoint to the desk tile, given
that officers were not able to reach the windoor previously.

Adds wall-mounted first aid cabinets with a roll of gauze and a pack
of ointment in Command, Engineering and Science. Also adds two first-aid
kits and a roller-bed in Security.

Fixes the access for the Trash Compactor.

Adds welding goggles to the welding lockers in Engineering, as well as
debuggers to the electrical lockers.

Adds a few missing air alarms and fire alarms in a few places.

Adds holsters to the Security clothing locker in the Brig.