Reworks Lichdom (#2292)
Instead of just turning the wizard into a skeleton with the resurrection proc, the lichdom spell will now create a phylactery, bound to the wizard. When the lich dies, he can use a skill to return to the phylactery, where he will come back after to life some time. If the phylactery is destroyed, the lich will be unable to return to life. Trying to use dark resurrection without a phylactery will result in the lich's body being dusted.

Also, adds a new sprite for the wizard book.