[Newmap] Modifications and bugfixes. (#2543)
Fixes: #2380, #2475, #2496, #2500, #2504, #2506, #2514, #2518, #2519.

Adds new flooring decals for the Medical Bay. These are of the same style that those that already existed,
but fit to the layout and naming of the new map's Medical Bay. There are new signs for the EMT, General Treatment Room, ICU, Ward and Cloning.(Images below)

Slightly modifies the layout of the Medbay, and the CE's office.
(Images below)