[New Map] Fixes Dumb Turrets (#2621)
Removes most instances of doublepane glass walls in one tile corridors, which the game was not designed for, thus interfering with turret target calculations and throwing objects. This means that the turrets in telecoms and in the vault will be more responsive to intruders.

Re-adds the fourth miner slot, but this time properly maps in a locker, spawn pos, hardsuit, and his own new type of flag - the purple flag (not that anyone uses flags ayy lmao.)

To further cement the uselessness of flags, all miners and the QM now start with a GPS.

CT's should now have access to the end of the miner belt, allowing them to collect resources.

Ore boxes can now be purchased and crafted from wood.