[New Map] Digging with a shovel no longer creates ore from nothing (#2611)
Using a shovel to dig a hole will now only spawn ore that that ore already holds. This means that an ore scanner is once again needed to maximise profits, although digging with a shovel remains far more inefficient than digging with a drill, due to simply how much sand there is in a sand tile.