Snowflake Qipao DLC (#2659)
Updates the red cheongsam to not be a stoopid recolor of the white one. Now it is a stupid recolor with a unique design. Attempts to make it show up over shoes have failed, to great sorrow. Recommend not wearing any cheongsams with jackboots.
Fixes quiche and keylimepie slice nutrition so that they don't go away in one bite.