Nerf Screenshaking (#2715)
Halves the max duration of screen shaking and adjusts the bottom line a bit. Previously, pretty much any explosion would have max intensity screen shaking because the minimum threshold was so high, and likewise for duration.

Standing right next to the epicenter of a sufficiently large explosion (I tested with size 16 bombs, the standard) will still cause some hefty screenshaking, as well as blindness and flash. However, standing at the very edge of an explosion, or standing next to a very small explosion's epicenter (such as an exploding durand or a frag grenade) will have only minor - but still notable (as all screenshaking is) screenshake.