Removes cloaking devices (#2713)
This pr will replace the cloaking device from cargo warehouse spawn,, because they pretty much fall under the same level of bullshit as thermals and etc, with the new portable ladders.

Also, removes from the traitor uplink, because it is only used for combat it seems, defeating the purpose of being a stealth item. The chameleon projector is a better, and far less bullshit item, to cover the same purpose.