Adjusts the tech levels for certain weapons (#2735)
This makes it possible to get the railgun and decloner eaiser. It's still rare and costly. Also made it possible to get the mech teleporter (It's really not as good as some people think) by using antag items to get bluespace research up to 6. Also removed a duplicate decloner fab.