Adds a docking location to the mercenary shuttle and more (#2767)
Adds a dock at arrivals for the mercenary shuttle to dock, also places the admin shuttle to a dock in the bridge.

Also, fixes the belt at robotics at last and adds a hull shield to engineering. And adds a hacked candy vending to the wizard base and ninja spawn, together with a suit cooling unit for ipc ninjas.