Sprite tweaks plus Ivancarp (#2845)
The top of the Tajaran industrial mining RIG chestpiece has been lowered to no-longer obscure the Tajaran's mouth from the front when the helmet is down. (Inconsistent with sidesprites anyway.)

The Ivancarp plushie can now spawn whenever the game calls for a random plushie, and the carp himself has a chance of appearing in cargo's warehouse.

Also made the latest attempt at fixing the phasing cigarettes.

Also changed that thing where part of the face is clearly exposed on the Rescue RIG's helmet when it is ''sealed''.