Fixes #2888 (#3038)
Since I'm sure people could come up with reasons any preset should have a file manager, I added it to all of them. Could undo it later, but it goes well with, say, the Relay program that lets you save (and view!) logs, as well as CERTAIN ROBOTICISTS (and mechatronic engineers) that like to put their forms on the computer.