Nukes aurorastation.int and silences invalid timer RTs during unit tests (#3105)
HACKS ARE NOT A SOLUTION TO THE PROBLEM.

True. But. The issue is convoluted to the point where fixing it is out of the question for the time being. And the invalid timer RT is ruining the readability of Travis tests: it's generating useless noise without conveying any new information. As such, until a time that this issue is fixed, I propose silencing the RTs generated by invalid timers for unit tests only.

Hopefully we won't fuck with the timer system enough to where we need to monitor it for mass invalid timer errors.

Also removes the .int file as the bug it was intended to fix has been fixed with BYOND.