[Newmap] Fixes and Security Equipment redesign.
- Remap: Security Office has been redesigned to make it more intuitive
and with a bit more of basic equipment.
- Addition: A new camera has been added to the Engine Bay, as AIs did
not have full view of the room (SMES and Emitter) when the shutters were
down.
- Fix: All docks and the mining airlock are now connected to Distro and
have a slightly different design. A small atmospheric substation has
been included, as well as a bypass, in case the Distro line is destroyed
and the docks don't have enough air to fill. Manually operated.
- Fix: Hopefully the issue causing the Departures shuttle not to dock
has been resolved.
- Fix: A suction vent has been installed in the Trash Compactor, to deal
with the increase of the pressure caused because of the disposals'
pneumatics.
- Fix: All powernet sensors have been toggled the long_range variable,
so they show up in the power screen.
- Fix: Access to the engine windoors has been fixed. Atmospheric
Technicians and Engineers have access.
- A revision of piping and wiring has been done, fixing improperly
connected pipes and power networks which mostly supplied electrified
windows.