Various coffee (and coffee maker) changes and fixes. (#3149)
Lets Cargo order coffee beans and a coffee maker, under Hospitality. The bridge, which is the only place that has a coffee maker, only has 100 units of black coffee and no way to replenish it. Most people don't even know about the hidden types of coffee, and I might add more in the future, so I wanted it to be more accessible.

Additionally, I made the two types of coffee actually able to be made.