Fixes Aurora3 issue 3498, which was a typo with the text when nanopaste is self used by an IPC and observed by another person.
There were 2 issues,
There was no space after a terniary function, causing a grammatical error.
Another grammatical error was caused due to the 2nd block of the terniary error being defined as " \the", which was processing as \t he (tab he) and causing results such as he Hephaetus instead of the Hephaetus.