Map & sprite fixes (#3755)
Fixes #3552

Also converts rest of all the blast doors on aurora to hopefully correct directions.
The standart should be NORTH and EAST. For window and multi-tile blast doors, they are always sideways to form a line, and when they cover a door they are always in the direction of travel.

Also fixes wrong camera direction in atmos sublevel room.

Also fixes APC sprite overlays not being properly rotated. Sorry for that one.

Also fixes those sideways stairs in CC cryo.