Added toggleable Autohiss for Dionaea (#3881)
The intent is to provide Dionaea players authentic sounding speech according to the currently existing lore. This PR does this by using the already existing autohiss plugin as well as a small custom sentence checker to change the sentence of whatever the player types. This only applies to non-rootsong languages and by default this is disabled.

When on basic or full, ellipses are added after the following words:
"who","what","when","where","why","how","i'm","i","am","this","they","are","they're","their","his","her","their","the","he","she"

When on basic or full, these letters are elongated:
"s","z","e"

When on full, these letters are elongated:
"a","i","o","u"

This was made with sleepy wolf's, the dionaea head loremaster, blessing.