Biogenerator Powercreep + Improvements (#4045)
The biogenerator seemed pretty useless on top of a generally boring job. There wasn't really anything interesting you could make with the default biogenerator recipes so I went on a quest to add more useful recipes that make sense. I looked at the code and it was pretty shit (everything was hardcoded, and there was also an exploit involving monkeys) so I added a better framework for adding biogenerator recipes, and added multiple item making.