[Ready for review until the inevitable map conflicts] Firing pins (#3375)
I'm going full steam ahead with this.

Adds firing pins/ Authentication mechanisms. these are needed to fire any specific weapon.

Weapons generally spawn with firing pins in them. or at least should. The exception to this is anything made in R&D - which needs to secure a firing pin from somewhere.

There's currently the following firing pins:

One that checks for a specific implant, EG a loyalty implant one.

One that always fails, and can replace other pins. (this one would likely be admin only. it's clown themed)

The Laser tag code got refactored to the pins. so each gun has a different pin. if you somehow got this pin, you could make it so you could only shoot a gun while wearing a laser tag vest, for example.

A Testing range pin, which only allows firing in specific areas - of which are mapped in currently.

A DNA locked one. and a subtype which will explode if someone else tries to use it.

the system's designed to be easy enough to add on to.

Feedback fourm is here: