[Ready for Review] Cane Upgrades for RP and Combat (#3965)
I found the cane to be annoying to use when it comes to doing actual work on the station. You can't use the cane to push buttons or generally interact with objects so I added something that would help.

But then I got carried away and added several other useful utility functions.

Using help on someone with the cane will poke them, dealing 0.25 pain damage that usually regenerates within 4 seconds. Spamming poke with the cane can cause brute damage (takes like 5 in a row for it to happen), and can actually knock people down (RNG, rare after 10 pokes). Useful for characters playing old disgruntled men who like hitting youngsters with the cane without killing them in 5 hits.
Using disarm on someone with the cane will perform an attack that takes 2.5 seconds to charge. If the arms are target, the weapon in that hand will be disarmed, else it will just deal 3 pain damage.
Using grab on someone with the cane will perform a pull that takes 2.5 seconds to charge. Equivalent to a normal Right Click > Pull however this one deals 1 pain damage on a successful pull.
Using harm on someone with the cane will perform a regular attack that deals 10 brute damage to the targeted body part.