Let's stack the FSM that is the reader. (#4369)
Okay.

So someone decided to strip the quotes from assoc lists. Which means they're no longer returned by default. So I stacked that condition onto the reader. Not sure if this is a good solution, but it works. The only thing to do other than this would be to redo the FSM to be a bit less stacked onto.